requirements.txt
Django==5.0.6
djangorestframework==3.15.1
numpy==2.1.1
python-dotenv==1.0.1
mysqlclient==2.2.4